refactor#97: 패키지 이름 변경
