+++ src/main/java/or/sopt/houme/domain/generateImage/entity/Type.java
@@ -1,0 +1,5 @@
+package or.sopt.houme.domain.generateImage.entity;
+
+public enum Type {
+    PNG, JPEG, WEBP, GIF
+}
