+++ src/test/java/or/sopt/houme/domain/user/repository/UserRepositoryImplTest.java
@@ -3,8 +3,8 @@ package or.sopt.houme.domain.user.repository;
 import jakarta.persistence.EntityManager;
 import or.sopt.houme.domain.credit.entity.Credit;
 import or.sopt.houme.domain.credit.entity.CreditStatus;
-import or.sopt.houme.domain.generatedImage.entity.GenerateImage;
-import or.sopt.houme.domain.generatedImage.entity.Type;
+import or.sopt.houme.domain.generateImage.entity.GenerateImage;
+import or.sopt.houme.domain.generateImage.entity.Type;
 import or.sopt.houme.domain.house.entity.House;
 import or.sopt.houme.domain.house.entity.enums.Activity;
 import or.sopt.houme.domain.house.entity.enums.Equilibrium;
